+++ jabgui/src/main/java/org/jabref/gui/fieldeditors/contextmenu/ContextAction.java
@@ -42,22 +42,12 @@ public class ContextAction extends SimpleCommand {
 
         this.executable.bind(
                 switch (command) {
-                    case RENAME_FILE_TO_PATTERN ->
+                    case RENAME_FILE_TO_PATTERN,
+                         RENAME_FILE_TO_NAME ->
                             Bindings.createBooleanBinding(
                                     () -> !linkedFile.getFile().isOnlineLink()
                                             && linkedFile.getFile().findIn(databaseContext, preferences.getFilePreferences()).isPresent()
                                             && !linkedFile.isGeneratedNameSameAsOriginal(),
-                                    nonNullDependencies(
-                                            linkedFile.getFile().linkProperty(),
-                                            entryFieldsObservable
-                                    )
-                            );
-
-                    case RENAME_FILE_TO_NAME ->
-                            Bindings.createBooleanBinding(
-                                    () -> !linkedFile.getFile().isOnlineLink()
-                                            && linkedFile.getFile().findIn(databaseContext, preferences.getFilePreferences()).isPresent()
-                                            && !linkedFile.isGeneratedPathSameAsOriginal(),
                                     nonNullDependencies(
                                             linkedFile.getFile().linkProperty(),
                                             entryFieldsObservable
@@ -84,8 +74,6 @@ public class ContextAction extends SimpleCommand {
 
                     case OPEN_FILE,
                          OPEN_FOLDER,
-                         MOVE_FILE_TO_FOLDER,
-                         MOVE_FILE_TO_FOLDER_AND_RENAME,
                          DELETE_FILE ->
                             Bindings.createBooleanBinding(
                                     () -> !linkedFile.getFile().isOnlineLink()
@@ -125,10 +113,6 @@ public class ContextAction extends SimpleCommand {
                     linkedFile.renameToSuggestion();
             case RENAME_FILE_TO_NAME ->
                     linkedFile.askForNameAndRename();
-            case MOVE_FILE_TO_FOLDER ->
-                    linkedFile.moveToNextConfiguredFileDirectory();
-            case MOVE_FILE_TO_FOLDER_AND_RENAME ->
-                    linkedFile.moveToNextConfiguredFileDirectoryAndRename();
             case DELETE_FILE ->
                     viewModel.deleteFile(linkedFile);
             case REMOVE_LINK,